Implemented room for database and save product without specification

+++ app/src/main/java/com/nurgazy_bolushbekov/product_informer/product/ProductSharedViewModel.kt
@@ -1,7 +1,9 @@
 package com.nurgazy_bolushbekov.product_informer.product
 
 import android.app.Application
+import android.graphics.Bitmap
 import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.nurgazy_bolushbekov.product_informer.data_classes.ProductResponse
 import com.nurgazy_bolushbekov.product_informer.product.image.ImageRepository
@@ -12,12 +14,13 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import java.io.File
 import javax.inject.Inject
 
 @HiltViewModel
-class ProductSharedViewModel @Inject constructor(application: Application): AndroidViewModel(application) {
-
-    private val imageRepository: ImageRepository = ImageRepositoryImpl(application)
+class ProductSharedViewModel @Inject constructor(
+    private val imageRepository: ImageRepositoryImpl,
+): ViewModel() {
 
     private val _currentProductResponse = MutableStateFlow<ProductResponse?>(null)
     val currentProductResponse: StateFlow<ProductResponse?> = _currentProductResponse.asStateFlow()
@@ -28,7 +31,8 @@ class ProductSharedViewModel @Inject constructor(application: Application): Andr
 
     fun deleteImage(){
         viewModelScope.launch(Dispatchers.IO) {
-            imageRepository.deleteImageFromCache(_currentProductResponse.value!!.savedImagePath!!)
+            if (_currentProductResponse.value!!.savedImagePath != null)
+                imageRepository.deleteImageFromCache(_currentProductResponse.value!!.savedImagePath!!)
         }
     }
 }
